+++ permissions/src/epilot/models/shared/__init__.py
@@ -11,4 +11,4 @@ from .security import *
 from .sharerole import *
 from .userrole import *
 
-__all__ = ["Assignment","Effect","Four","Grant","GrantCondition","One","Operation","OrgRole","PartnerRole","PartnerRoleType","RolePayload","RolePayloadSchemasTrueRolePayloadType","RolePayloadSchemasTrueType","RolePayloadSchemasType","RolePayloadType","RoleSearchInput","Schemas","Security","ShareRole","ShareRoleType","Three","Two","Type","UserRole","UserRoleType"]
+__all__ = ["Assignment","Effect","Four","Grant","One","Operation","OrgRole","PartnerRole","PartnerRoleType","RolePayloadSchemasTrueRolePayloadType","RolePayloadSchemasTrueType","RolePayloadSchemasType","RolePayloadType","RoleSearchInput","Schemas","Security","ShareRole","ShareRoleType","Three","Two","Type","UserRole","UserRoleType"]