+++ permissions/src/epilot/models/shared/grantcondition.py
@@ -20,8 +20,3 @@ class Schemas:
     values: List[Any] = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('values') }})
     
 
-
-
-@dataclasses.dataclass
-class GrantCondition:
-    pass